Move `IoTypeMetadataKind` and `concat_metadata_sources` into `metadata` module
