+++ crates/contracts/ab-contracts-macros/src/__private.rs
@@ -2,6 +2,6 @@ pub use ab_contracts_common::env::{Env, MethodContext};
 pub use ab_contracts_common::metadata::ContractMetadataKind;
 pub use ab_contracts_common::methods::{ExternalArgs, MethodFingerprint};
 pub use ab_contracts_common::{Address, ContractError, ExitCode};
+pub use ab_contracts_io_type::metadata::concat_metadata_sources;
 pub use ab_contracts_io_type::trivial_type::TrivialType;
-pub use ab_contracts_io_type::utils::concat_metadata_sources;
 pub use ab_contracts_io_type::{IoType, IoTypeOptional};
