+++ crates/contracts/ab-contracts-trivial-type-derive/src/lib.rs
@@ -280,8 +280,8 @@ fn generate_struct_metadata(ident: &Ident, data_struct: &DataStruct) -> Result<T
     // * The rest as inner struct metadata
     Ok(quote! {{
         const fn metadata() -> ([u8; 4096], usize) {
-            ::ab_contracts_io_type::utils::concat_metadata_sources(&[
-                &[::ab_contracts_io_type::IoTypeMetadataKind::#io_type_metadata as u8],
+            ::ab_contracts_io_type::metadata::concat_metadata_sources(&[
+                &[::ab_contracts_io_type::metadata::IoTypeMetadataKind::#io_type_metadata as u8],
                 #( #inner_struct_metadata )*
             ])
         }
@@ -346,7 +346,7 @@ fn generate_enum_metadata(ident: &Ident, data_enum: &DataEnum) -> Result<TokenSt
 
         quote! {
             &[
-                ::ab_contracts_io_type::IoTypeMetadataKind::#io_type_metadata as u8,
+                ::ab_contracts_io_type::metadata::IoTypeMetadataKind::#io_type_metadata as u8,
                 #( #enum_metadata_header, )*
             ]
         }
@@ -361,7 +361,7 @@ fn generate_enum_metadata(ident: &Ident, data_enum: &DataEnum) -> Result<TokenSt
 
     Ok(quote! {{
         const fn metadata() -> ([u8; 4096], usize) {
-            ::ab_contracts_io_type::utils::concat_metadata_sources(&[
+            ::ab_contracts_io_type::metadata::concat_metadata_sources(&[
                 #enum_metadata_header,
                 #( #inner )*
             ])
